move logger to its own directory
