+++ use-cases/save-pokemons-in-csv.go
@@ -2,7 +2,7 @@ package usecases
 
 import (
 	"github.com/georgexx009-wizeline/ondemand-go-bootcamp/model"
-	logger "github.com/georgexx009-wizeline/ondemand-go-bootcamp/pkg"
+	"github.com/georgexx009-wizeline/ondemand-go-bootcamp/pkg/logger"
 	"reflect"
 	"strconv"
 )
